requirements.txt
requests
beautifulsoup4
rich
feedparser
html2text
schedule
openai
python-dotenv
